Update HelloController.java
I inputted code for each graph based on their fxids. I added arrays for each of the data series. To convert the two arrays to a series to be plotted on the graph, I used a for loop and a nested for loop. 
Please feel free to test out these changes and make your own changes!
Also, for anyone who needs info on the graph code, I added a notes section under the controller class that has a messy but step-by-step explanation.